Enable publishing from anywhere.
Previously publishing was only capable internally because
Microsoft.IIS.Host was not available. Now Microsoft.IIS.Host is obtained
through NuGet during the publish process.

The installation was updated to use the new directory structure of the
Microsoft.IIS.Host package. x64\win32 became win32\x64. The services
script had to be updated to reflect these changes.

A bug was fixed in the migrate script which caused invalid format in the
modules.json file.