Added checks for allowed paths to compression, logging, and request tracing when changing the directory for storing files.
Fixed an error in feature-delegation that was caused when a configuration
file contained a locked section.

Sites, apps, and vdirs now check for non-normalized physical paths ('..',
and '.' in the path), and return 400 (Bad Request) if such a path is
given.

Sites, apps, and vdirs now return 404 (Not Found) when the directory specified by
physical_path does not exist, rather than 400 (Bad Request).

Api Explorer now supports range uploads for large files.

Api Explorer now shows the upload button only when the "PUT" verb is
selected.